requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
openai==1.47.0
pandas==2.2.2
plotly==5.24.0
orjson==3.10.7
streamlit==1.38.0
requests==2.32.3
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
langchain-core==0.3.9